Code/backend/requirements.txt
Flask
Flask-CORS
SQLAlchemy
python-dotenv
psycopg2-binary
transformers
torch
langdetect
flask
flask-cors
flask_sqlalchemy
textblob